Add shared component for `SingleMessageFieldTypesProvider`.
